Fixes for Visual Studio 2017.

+++ src/gui/texture_font.cpp
@@ -38,7 +38,7 @@ namespace Yttrium
 			const float scaling = font_size / _size;
 
 			float selection_left = 0;
-			const auto do_capture = [font_size, capture, &current_x, current_y, &selection_left](unsigned index)
+			const auto do_capture = [font_size, capture, &current_x, current_y, &selection_left](size_t index)
 			{
 				if (!capture)
 					return;
